Miner/requirements.txt
appnope==0.1.3
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
brotlipy==0.7.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.5
cryptography==37.0.1
debugpy==1.6.0
decorator==5.1.1
Deprecated==1.2.13
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.16.1
firebase-admin==5.2.0
gitdb==4.0.9
GitPython==3.1.27
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.1
google-cloud-firestore==2.5.3
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
ipykernel==6.13.0
ipython==8.2.0
javalang==0.13.0
jedi==0.18.1
jsonschema==4.7.2
jupyter-client==7.2.2
jupyter-core==4.10.0
matplotlib-inline==0.1.3
msgpack==1.0.4
nbformat==5.4.0
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pigar==1.0.0
prompt-toolkit==3.0.29
proto-plus==1.20.6
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.4.0
pylibsrcml==1.0.0.post1
PyNaCl==1.5.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.28.1
rsa==4.8
six==1.16.0
smmap==5.0.0
stack-data==0.2.0
tornado==6.1
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.10
wcwidth==0.2.5
wrapt==1.14.1
